remove unused gitlab api package
