+++ internal/data/mergerequest.mock.go
@@ -4,284 +4,194 @@ import (
 	"time"
 
 	"github.com/felipeospina21/mrglab/internal/gql"
-	"github.com/xanzy/go-gitlab"
 )
 
-var (
-	MockUser          = &gitlab.BasicUser{Name: "Dummy"}
-	MergerequestsMock = []*gitlab.MergeRequest{
+var GQLMergeRequestMock = gql.MergeRequestConnection{
+	Count: 1,
+	Edges: []gql.MergeRequestEdge{
 		{
-			IID:                 1,
-			Title:               "Mocked Title",
-			CreatedAt:           &time.Time{},
-			Draft:               true,
-			Author:              MockUser,
-			DetailedMergeStatus: "draft_status",
-			HasConflicts:        false,
-			UserNotesCount:      4,
-			WebURL:              "",
-			Description:         "Dummy description",
-		},
-		{
-			IID:                 2,
-			Title:               "Mocked Title 2",
-			CreatedAt:           &time.Time{},
-			Draft:               false,
-			Author:              MockUser,
-			DetailedMergeStatus: "conflict",
-			HasConflicts:        true,
-			UserNotesCount:      0,
-			WebURL:              "",
-			Description:         "Dummy description",
-		},
-		{
-			IID:                 3,
-			Title:               "Mocked Title 3",
-			CreatedAt:           &time.Time{},
-			Draft:               false,
-			Author:              MockUser,
-			DetailedMergeStatus: "need_rebase",
-			HasConflicts:        true,
-			UserNotesCount:      0,
-			WebURL:              "",
-			Description:         "Dummy description",
+			Node: gql.MergeRequestNode{
+				IID:                 "1",
+				Title:               "Mocked Title 1",
+				CreatedAt:           time.Now(),
+				UpdatedAt:           time.Now(),
+				Draft:               false,
+				Author:              gql.Author{Name: "Mock User"},
+				DetailedMergeStatus: "ci_still_running",
+				Conflicts:           true,
+				UserNotesCount:      0,
+				WebURL:              "",
+				Description:         "Dummy description",
+				ApprovalsLeft:       3,
+				ApprovalsRequired:   3,
+				DiffStatsSummary: gql.DiffStatsSummary{
+					Additions: 10,
+					Deletions: 50,
+					Changes:   100,
+					FileCount: 5,
+				},
+			},
 		},
 		{
-			IID:                 4,
-			Title:               "Mocked Title 4",
-			CreatedAt:           &time.Time{},
-			Draft:               true,
-			Author:              MockUser,
-			DetailedMergeStatus: "mergeable",
-			HasConflicts:        false,
-			UserNotesCount:      4,
-			WebURL:              "",
-			Description:         "Dummy description",
+			Node: gql.MergeRequestNode{
+				IID:                 "2",
+				Title:               "Mocked Title 2",
+				CreatedAt:           time.Now(),
+				UpdatedAt:           time.Now(),
+				Draft:               false,
+				Author:              gql.Author{Name: "Mock User"},
+				DetailedMergeStatus: "blocked_status",
+				Conflicts:           true,
+				UserNotesCount:      10,
+				WebURL:              "",
+				Description:         "Dummy description",
+				ApprovalsLeft:       3,
+				ApprovalsRequired:   3,
+				DiffStatsSummary: gql.DiffStatsSummary{
+					Additions: 10,
+					Deletions: 50,
+					Changes:   100,
+					FileCount: 5,
+				},
+			},
 		},
 		{
-			IID:                 5,
-			Title:               "Mocked Title 5",
-			CreatedAt:           &time.Time{},
-			Draft:               false,
-			Author:              MockUser,
-			DetailedMergeStatus: "blocked_status",
-			HasConflicts:        true,
-			UserNotesCount:      0,
-			WebURL:              "",
-			Description:         "Dummy description",
+			Node: gql.MergeRequestNode{
+				IID:                 "3",
+				Title:               "Mocked Title 3",
+				CreatedAt:           time.Now(),
+				UpdatedAt:           time.Now(),
+				Draft:               false,
+				Author:              gql.Author{Name: "Mock User"},
+				DetailedMergeStatus: "conflicts",
+				Conflicts:           true,
+				UserNotesCount:      0,
+				WebURL:              "",
+				Description:         "Dummy description",
+				ApprovalsLeft:       3,
+				ApprovalsRequired:   3,
+				DiffStatsSummary: gql.DiffStatsSummary{
+					Additions: 10,
+					Deletions: 50,
+					Changes:   100,
+					FileCount: 5,
+				},
+			},
 		},
 		{
-			IID:                 6,
-			Title:               "Mocked Title 6",
-			CreatedAt:           &time.Time{},
-			Draft:               false,
-			Author:              MockUser,
-			DetailedMergeStatus: "discussions_not_resolved",
-			HasConflicts:        true,
-			UserNotesCount:      0,
-			WebURL:              "",
-			Description:         "Dummy description",
+			Node: gql.MergeRequestNode{
+				IID:                 "4",
+				Title:               "Mocked Title 4",
+				CreatedAt:           time.Now(),
+				UpdatedAt:           time.Now(),
+				Draft:               false,
+				Author:              gql.Author{Name: "Mock User"},
+				DetailedMergeStatus: "mergeable",
+				Conflicts:           false,
+				UserNotesCount:      0,
+				WebURL:              "",
+				Description:         "Dummy description",
+				ApprovalsLeft:       3,
+				ApprovalsRequired:   3,
+				DiffStatsSummary: gql.DiffStatsSummary{
+					Additions: 10,
+					Deletions: 50,
+					Changes:   100,
+					FileCount: 5,
+				},
+			},
 		},
 		{
-			IID:                 7,
-			Title:               "Mocked Title 7",
-			CreatedAt:           &time.Time{},
-			Draft:               false,
-			Author:              MockUser,
-			DetailedMergeStatus: "ci_still_running",
-			HasConflicts:        true,
-			UserNotesCount:      0,
-			WebURL:              "",
-			Description:         "Dummy description",
-		},
-	}
-	GQLMergeRequestMock = gql.MergeRequestConnection{
-		Count: 1,
-		Edges: []gql.MergeRequestEdge{
-			{
-				Node: gql.MergeRequestNode{
-					IID:                 "1",
-					Title:               "Mocked Title 1",
-					CreatedAt:           time.Now(),
-					UpdatedAt:           time.Now(),
-					Draft:               false,
-					Author:              gql.Author{Name: "Mock User"},
-					DetailedMergeStatus: "ci_still_running",
-					Conflicts:           true,
-					UserNotesCount:      0,
-					WebURL:              "",
-					Description:         "Dummy description",
-					ApprovalsLeft:       3,
-					ApprovalsRequired:   3,
-					DiffStatsSummary: gql.DiffStatsSummary{
-						Additions: 10,
-						Deletions: 50,
-						Changes:   100,
-						FileCount: 5,
-					},
-				},
-			},
-			{
-				Node: gql.MergeRequestNode{
-					IID:                 "2",
-					Title:               "Mocked Title 2",
-					CreatedAt:           time.Now(),
-					UpdatedAt:           time.Now(),
-					Draft:               false,
-					Author:              gql.Author{Name: "Mock User"},
-					DetailedMergeStatus: "blocked_status",
-					Conflicts:           true,
-					UserNotesCount:      10,
-					WebURL:              "",
-					Description:         "Dummy description",
-					ApprovalsLeft:       3,
-					ApprovalsRequired:   3,
-					DiffStatsSummary: gql.DiffStatsSummary{
-						Additions: 10,
-						Deletions: 50,
-						Changes:   100,
-						FileCount: 5,
-					},
-				},
-			},
-			{
-				Node: gql.MergeRequestNode{
-					IID:                 "3",
-					Title:               "Mocked Title 3",
-					CreatedAt:           time.Now(),
-					UpdatedAt:           time.Now(),
-					Draft:               false,
-					Author:              gql.Author{Name: "Mock User"},
-					DetailedMergeStatus: "conflicts",
-					Conflicts:           true,
-					UserNotesCount:      0,
-					WebURL:              "",
-					Description:         "Dummy description",
-					ApprovalsLeft:       3,
-					ApprovalsRequired:   3,
-					DiffStatsSummary: gql.DiffStatsSummary{
-						Additions: 10,
-						Deletions: 50,
-						Changes:   100,
-						FileCount: 5,
-					},
-				},
-			},
-			{
-				Node: gql.MergeRequestNode{
-					IID:                 "4",
-					Title:               "Mocked Title 4",
-					CreatedAt:           time.Now(),
-					UpdatedAt:           time.Now(),
-					Draft:               false,
-					Author:              gql.Author{Name: "Mock User"},
-					DetailedMergeStatus: "mergeable",
-					Conflicts:           false,
-					UserNotesCount:      0,
-					WebURL:              "",
-					Description:         "Dummy description",
-					ApprovalsLeft:       3,
-					ApprovalsRequired:   3,
-					DiffStatsSummary: gql.DiffStatsSummary{
-						Additions: 10,
-						Deletions: 50,
-						Changes:   100,
-						FileCount: 5,
-					},
-				},
-			},
-			{
-				Node: gql.MergeRequestNode{
-					IID:                 "5",
-					Title:               "Mocked Title 5",
-					CreatedAt:           time.Now(),
-					UpdatedAt:           time.Now(),
-					Draft:               false,
-					Author:              gql.Author{Name: "Mock User"},
-					DetailedMergeStatus: "external_status_checks",
-					Conflicts:           false,
-					UserNotesCount:      0,
-					WebURL:              "",
-					Description:         "Dummy description",
-					ApprovalsLeft:       3,
-					ApprovalsRequired:   3,
-					DiffStatsSummary: gql.DiffStatsSummary{
-						Additions: 10,
-						Deletions: 50,
-						Changes:   100,
-						FileCount: 5,
-					},
-				},
-			},
-			{
-				Node: gql.MergeRequestNode{
-					IID:                 "6",
-					Title:               "Mocked Title 6",
-					CreatedAt:           time.Now(),
-					UpdatedAt:           time.Now(),
-					Draft:               false,
-					Author:              gql.Author{Name: "Mock User"},
-					DetailedMergeStatus: "need_rebase",
-					Conflicts:           false,
-					UserNotesCount:      0,
-					WebURL:              "",
-					Description:         "Dummy description",
-					ApprovalsLeft:       3,
-					ApprovalsRequired:   3,
-					DiffStatsSummary: gql.DiffStatsSummary{
-						Additions: 10,
-						Deletions: 50,
-						Changes:   100,
-						FileCount: 5,
-					},
-				},
-			},
-			{
-				Node: gql.MergeRequestNode{
-					IID:                 "7",
-					Title:               "Mocked Title 7",
-					CreatedAt:           time.Now(),
-					UpdatedAt:           time.Now(),
-					Draft:               false,
-					Author:              gql.Author{Name: "Mock User"},
-					DetailedMergeStatus: "unchecked",
-					Conflicts:           false,
-					UserNotesCount:      0,
-					WebURL:              "",
-					Description:         "Dummy description",
-					ApprovalsLeft:       1,
-					ApprovalsRequired:   3,
-					DiffStatsSummary: gql.DiffStatsSummary{
-						Additions: 10,
-						Deletions: 50,
-						Changes:   100,
-						FileCount: 5,
-					},
-				},
-			},
-			{
-				Node: gql.MergeRequestNode{
-					IID:                 "8",
-					Title:               "Mocked Title 8",
-					CreatedAt:           time.Now(),
-					UpdatedAt:           time.Now(),
-					Draft:               false,
-					Author:              gql.Author{Name: "Mock User"},
-					DetailedMergeStatus: "ci_still_running",
-					Conflicts:           false,
-					UserNotesCount:      0,
-					WebURL:              "",
-					Description:         "Dummy description",
-					ApprovalsLeft:       2,
-					ApprovalsRequired:   3,
-					DiffStatsSummary: gql.DiffStatsSummary{
-						Additions: 10,
-						Deletions: 50,
-						Changes:   100,
-						FileCount: 5,
-					},
+			Node: gql.MergeRequestNode{
+				IID:                 "5",
+				Title:               "Mocked Title 5",
+				CreatedAt:           time.Now(),
+				UpdatedAt:           time.Now(),
+				Draft:               false,
+				Author:              gql.Author{Name: "Mock User"},
+				DetailedMergeStatus: "external_status_checks",
+				Conflicts:           false,
+				UserNotesCount:      0,
+				WebURL:              "",
+				Description:         "Dummy description",
+				ApprovalsLeft:       3,
+				ApprovalsRequired:   3,
+				DiffStatsSummary: gql.DiffStatsSummary{
+					Additions: 10,
+					Deletions: 50,
+					Changes:   100,
+					FileCount: 5,
 				},
 			},
 		},
-	}
-)
+		{
+			Node: gql.MergeRequestNode{
+				IID:                 "6",
+				Title:               "Mocked Title 6",
+				CreatedAt:           time.Now(),
+				UpdatedAt:           time.Now(),
+				Draft:               false,
+				Author:              gql.Author{Name: "Mock User"},
+				DetailedMergeStatus: "need_rebase",
+				Conflicts:           false,
+				UserNotesCount:      0,
+				WebURL:              "",
+				Description:         "Dummy description",
+				ApprovalsLeft:       3,
+				ApprovalsRequired:   3,
+				DiffStatsSummary: gql.DiffStatsSummary{
+					Additions: 10,
+					Deletions: 50,
+					Changes:   100,
+					FileCount: 5,
+				},
+			},
+		},
+		{
+			Node: gql.MergeRequestNode{
+				IID:                 "7",
+				Title:               "Mocked Title 7",
+				CreatedAt:           time.Now(),
+				UpdatedAt:           time.Now(),
+				Draft:               false,
+				Author:              gql.Author{Name: "Mock User"},
+				DetailedMergeStatus: "unchecked",
+				Conflicts:           false,
+				UserNotesCount:      0,
+				WebURL:              "",
+				Description:         "Dummy description",
+				ApprovalsLeft:       1,
+				ApprovalsRequired:   3,
+				DiffStatsSummary: gql.DiffStatsSummary{
+					Additions: 10,
+					Deletions: 50,
+					Changes:   100,
+					FileCount: 5,
+				},
+			},
+		},
+		{
+			Node: gql.MergeRequestNode{
+				IID:                 "8",
+				Title:               "Mocked Title 8",
+				CreatedAt:           time.Now(),
+				UpdatedAt:           time.Now(),
+				Draft:               false,
+				Author:              gql.Author{Name: "Mock User"},
+				DetailedMergeStatus: "ci_still_running",
+				Conflicts:           false,
+				UserNotesCount:      0,
+				WebURL:              "",
+				Description:         "Dummy description",
+				ApprovalsLeft:       2,
+				ApprovalsRequired:   3,
+				DiffStatsSummary: gql.DiffStatsSummary{
+					Additions: 10,
+					Deletions: 50,
+					Changes:   100,
+					FileCount: 5,
+				},
+			},
+		},
+	},
+}
